LPD-48502 Move
