+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/selectors/selectStructureField.ts
@@ -3,7 +3,8 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {Field, State} from '../contexts/StateContext';
+import {State} from '../contexts/StateContext';
+import {Field} from '../utils/field';
 
 export default function selectStructureField(name: Field['name']) {
 	return (state: State) => {
