+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/services/StructureService.ts
@@ -3,8 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {Field, State} from '../contexts/StateContext';
+import {State} from '../contexts/StateContext';
 import buildObjectDefinition from '../utils/buildObjectDefinition';
+import {Field} from '../utils/field';
 import ApiHelper from './ApiHelper';
 
 async function createStructure({
